requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-inmemorystorage==2.1.0
Django==3.0.7
django-allauth==0.42.0
django-environ==0.4.5
djangorestframework==3.11.0
drf-action-serializer==1.0.5
drf-nested-routers==0.91
factory-boy==2.12.0
Faker==4.1.1
idna==2.10
model-bakery==1.1.0
munch==2.5.0
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
